Rearrange the `rand` features to support both 0.9 and 0.10
