+++ Cargo.toml
@@ -19,25 +19,30 @@ default = ["std"]
 std = ["num-integer/std", "num-traits/std"]
 arbitrary = ["dep:arbitrary"]
 quickcheck = ["dep:quickcheck"]
-rand = ["rand_core", "dep:rand"]
-rand_core = ["dep:rand_core"]
+rand_0_9 = ["rand_core_0_9", "dep:rand_0_9"]
+rand_core_0_9 = ["dep:rand_core_0_9"]
+rand_0_10 = ["rand_core_0_10", "dep:rand_0_10"]
+rand_core_0_10 = ["dep:rand_core_0_10"]
 serde = ["dep:serde"]
 
 [package.metadata.docs.rs]
-features = ["std", "serde", "rand", "quickcheck", "arbitrary"]
+features = ["std", "serde", "rand_0_10", "rand_0_9", "quickcheck", "arbitrary"]
 rustdoc-args = ["--cfg", "docsrs"]
 
 [[bench]]
 name = "bigint"
+required-features = ["rand_0_10"]
 
 [[bench]]
 name = "factorial"
 
 [[bench]]
 name = "gcd"
+required-features = ["rand_0_10"]
 
 [[bench]]
 name = "roots"
+required-features = ["rand_0_10"]
 
 [[bench]]
 harness = false
@@ -55,12 +60,26 @@ version = "0.2.18"
 default-features = false
 features = ["i128"]
 
-[dependencies.rand]
+[dependencies.rand_0_9]
+package = "rand"
 optional = true
-version = "0.10.0-rc.9"
+version = "0.9"
 default-features = false
 
-[dependencies.rand_core]
+[dependencies.rand_core_0_9]
+package = "rand_core"
+optional = true
+version = "0.9"
+default-features = false
+
+[dependencies.rand_0_10]
+package = "rand"
+optional = true
+version = "0.10"
+default-features = false
+
+[dependencies.rand_core_0_10]
+package = "rand_core"
 optional = true
 version = "0.10"
 default-features = false
