+++ ci/big_rand_0_10/Cargo.toml
@@ -1,17 +1,16 @@
 [package]
 name = "big_rand"
-version = "0.1.0"
-authors = ["Josh Stone <cuviper@gmail.com>"]
-edition = "2018"
+edition = "2024"
 rust-version = "1.85"
+publish = false
 
 [dependencies]
 num-traits = "0.2.11"
-rand = "0.10.0-rc.9"
+rand = "0.10"
 rand_chacha = "0.10"
 rand_isaac = "0.5"
 rand_xorshift = "0.5"
 
 [dependencies.num-bigint]
-features = ["rand"]
+features = ["rand_0_10"]
 path = "../.."
