+++ ci/big_rand_0_10/src/torture.rs
@@ -1,4 +1,5 @@
-use num_bigint::RandBigInt;
+use crate::BigRng;
+
 use num_traits::Zero;
 use rand::prelude::*;
 use rand_xorshift::XorShiftRng;
